Add post hook and check if certificate changed before downloading

+++ Program.cs
@@ -5,6 +5,7 @@ using Mono.Unix;
 using Serilog;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography.X509Certificates;
@@ -67,16 +68,47 @@ namespace keyvault_certsync
                         return -1;
                     }
 
-                    return DownloadCertificate(opts, client, cert);
+                    if (!opts.Quiet)
+                    {
+                        Log.Information("Processing certifiate {Name}", cert.CertificateName);
+                        Console.WriteLine(cert.ToString());
+                    }
+
+                    var result = DownloadCertificate(opts, client, cert);
+
+                    if (result == DownloadResult.Error)
+                        return -1;
+
+                    if(result == DownloadResult.Success)
+                        return RunPostHook(opts);
+
+                    return 0;
                 }
                 else if (opts.Download)
-                { 
-                    int result = 0;
+                {
+                    bool downloaded = false;
+                    bool error = false;
                     foreach (var cert in client.GetCertificateDetails())
-                        if (DownloadCertificate(opts, client, cert) < 0)
-                            result = -1;
+                    {
+                        if (!opts.Quiet)
+                        {
+                            Log.Information("Processing certifiate {Name}", cert.CertificateName);
+                            Console.WriteLine(cert.ToString());
+                        }
 
-                    return result;
+                        var result = DownloadCertificate(opts, client, cert);
+
+                        if (result == DownloadResult.Error)
+                            error = true;
+
+                        if (result == DownloadResult.Success)
+                            downloaded = true;
+                    }
+
+                    if(downloaded)
+                        return RunPostHook(opts);
+
+                    return error ? -1 : 0;
                 }
             }
             catch (CredentialUnavailableException ex)
@@ -98,7 +130,52 @@ namespace keyvault_certsync
             return -1;
         }
 
-        private static int DownloadCertificate(Options opts, SecretClient client, CertificateDetails cert)
+        private static bool IdenticalLocalCertificatePath(Options opts, CertificateDetails cert)
+        {
+            if (!File.Exists(cert.GetPath(opts.Path, CERT_PEM)))
+                return false;
+
+            X509Certificate2 x509;
+            try
+            {
+                x509 = new X509Certificate2(cert.GetPath(opts.Path, CERT_PEM));
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, "Error reading local certificate");
+                return false;
+            }
+
+            if (string.Equals(cert.Thumbprint, x509.Thumbprint, StringComparison.CurrentCultureIgnoreCase))
+            {
+                if (!opts.Quiet)
+                    Log.Information("Local certificate has identical thumbprint");
+
+                return true;
+            }
+
+            return false;
+        }
+
+        private static bool IdenticalLocalCertificateStore(Options opts, CertificateDetails cert)
+        {
+            using X509Store store = new X509Store(opts.Store.Value);
+            store.Open(OpenFlags.ReadOnly);
+
+            var found = store.Certificates.Find(X509FindType.FindByThumbprint, cert.Thumbprint, false);
+
+            if (found.Count > 0 && found[0].HasPrivateKey)
+            {
+                if (!opts.Quiet)
+                    Log.Information("Local certificate has identical thumbprint");
+
+                return true;
+            }
+
+            return false;
+        }
+
+        private static DownloadResult DownloadCertificate(Options opts, SecretClient client, CertificateDetails cert)
         {
             X509Certificate2Collection chain;
             try
@@ -108,57 +185,64 @@ namespace keyvault_certsync
             catch (Azure.RequestFailedException ex)
             {
                 Log.Error(ex, "Error downloading certificate from Key Vault");
-                return -1;
+                return DownloadResult.Error;
             }
             catch (NotSupportedException ex)
             {
                 Log.Error(ex, "Key Vault certificate is invalid");
-                return -1;
+                return DownloadResult.Error;
             }
 
             if (!string.IsNullOrEmpty(opts.Path))
             {
-                DownloadCertificatePath(opts, cert, chain);
-                return 0;
+                return DownloadCertificatePath(opts, cert, chain);
             }
             else if (opts.Store.HasValue)
             {
                 if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
                 {
                     Log.Error("Certificate store is only supported on Windows");
-                    return -1;
+                    return DownloadResult.Error;
                 }
 
-                DownloadCertificateStore(opts, cert, chain);
-                return 0;
+                return DownloadCertificateStore(opts, cert, chain);
             }
 
-            return 0;
+            return DownloadResult.Success;
         }
 
-        private static void DownloadCertificatePath(Options opts, CertificateDetails cert, X509Certificate2Collection chain)
+        private static DownloadResult DownloadCertificatePath(Options opts, CertificateDetails cert, X509Certificate2Collection chain)
         {
             if (!Directory.Exists(cert.GetPath(opts.Path)))
                 Directory.CreateDirectory(cert.GetPath(opts.Path));
-            
-            if(!opts.Quiet)
+
+            if (!opts.Force && IdenticalLocalCertificatePath(opts, cert))
+                return DownloadResult.AlreadyExists;
+
+            if (!opts.Quiet)
                 Log.Information("Downloading certificate to {Path}", cert.GetPath(opts.Path));
 
-            StringBuilder pemFulChain = new StringBuilder();
+            StringBuilder pemFullChain = new StringBuilder();
+
+            if (!opts.Quiet)
+                Log.Information("Adding certificate {Subject}", chain[0].Subject);
 
             string pemCert = chain[0].ToCertificatePEM();
-            pemFulChain.AppendLine(pemCert);
+            pemFullChain.AppendLine(pemCert);
             File.WriteAllText(cert.GetPath(opts.Path, CERT_PEM), pemCert);
 
             StringBuilder pemChain = new StringBuilder();
             for (int i = 1; i < chain.Count; i++)
             {
+                if (!opts.Quiet)
+                    Log.Information("Adding chain certificate {Subject}", chain[i].Subject);
+
                 pemChain.AppendLine(chain[i].ToCertificatePEM());
-                pemFulChain.AppendLine(chain[i].ToCertificatePEM());
+                pemFullChain.AppendLine(chain[i].ToCertificatePEM());
             }
 
             File.WriteAllText(cert.GetPath(opts.Path, CHAIN_PEM), pemChain.ToString());
-            File.WriteAllText(cert.GetPath(opts.Path, FULLCHAIN_PEM), pemFulChain.ToString());
+            File.WriteAllText(cert.GetPath(opts.Path, FULLCHAIN_PEM), pemFullChain.ToString());
 
             if (chain[0].HasPrivateKey)
             {
@@ -167,11 +251,14 @@ namespace keyvault_certsync
                 CreateFileWithUserReadWrite(cert.GetPath(opts.Path, PRIVKEY_PEM));
                 File.WriteAllText(cert.GetPath(opts.Path, PRIVKEY_PEM), privKey);
 
-                pemFulChain.AppendLine(privKey);
+                pemFullChain.AppendLine(privKey);
                 CreateFileWithUserReadWrite(cert.GetPath(opts.Path, FULLKEYCHAIN_PEM));
-                File.WriteAllText(cert.GetPath(opts.Path, FULLKEYCHAIN_PEM), pemFulChain.ToString());
+                File.WriteAllText(cert.GetPath(opts.Path, FULLKEYCHAIN_PEM), pemFullChain.ToString());
             }
+
+            return DownloadResult.Success;
         }
+
         private static void CreateFileWithUserReadWrite(string filename)
         {
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
@@ -182,10 +269,16 @@ namespace keyvault_certsync
             }
         }
 
-        private static void DownloadCertificateStore(Options opts, CertificateDetails cert, X509Certificate2Collection chain)
+        private static DownloadResult DownloadCertificateStore(Options opts, CertificateDetails cert, X509Certificate2Collection chain)
         {
+            if (!opts.Force && IdenticalLocalCertificateStore(opts, cert))
+                return DownloadResult.AlreadyExists;
+
             using X509Store store = new X509Store(opts.Store.Value);
             store.Open(OpenFlags.ReadWrite);
+
+            if (!opts.Quiet)
+                Log.Information("Adding certificate {Subject}", chain[0].Subject);
 
             chain[0].FriendlyName = cert.CertificateName;
             store.Add(chain[0]);
@@ -194,7 +287,54 @@ namespace keyvault_certsync
             rootStore.Open(OpenFlags.ReadWrite);
 
             for (int i = 1; i < chain.Count; i++)
+            {
+                if (!opts.Quiet)
+                    Log.Information("Adding chain certificate {Subject}", chain[i].Subject);
+
                 rootStore.Add(chain[i]);
+            }
+
+            return DownloadResult.Success;
+        }
+
+        private static int RunPostHook(Options opts)
+        {
+            if (string.IsNullOrEmpty(opts.PostHook))
+                return 0;
+
+            string[] parts = opts.PostHook.Split(new[] { ' ' }, 2);
+
+            ProcessStartInfo startInfo;
+
+            if (parts.Length > 1)
+                startInfo = new ProcessStartInfo()
+                {
+                    FileName = parts[0],
+                    Arguments = parts[1]
+                };
+            else
+                startInfo = new ProcessStartInfo(parts[0]);
+
+            int exitCode;
+            try
+            {
+
+                using var process = Process.Start(startInfo);
+                process.WaitForExit();
+                exitCode = process.ExitCode;
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, "Unable to start post hook '{Hook}' '{HookArguments}'", startInfo.FileName, startInfo.Arguments);
+                return -1;
+            }
+
+            if (!opts.Quiet && exitCode == 0)
+                Log.Information("Post hook '{Hook}' '{HookArguments}' completed successfully", startInfo.FileName, startInfo.Arguments);
+            else if (!opts.Quiet)
+                Log.Warning("Post hook '{Hook}' '{HookArguments}' completed with exit code {ExitCode}", startInfo.FileName, startInfo.Arguments, exitCode);
+
+            return exitCode != 0 ? exitCode : 0;
         }
 
         private static int HandleParseError(IEnumerable<Error> errs)
